연결 오류 일부 해결

+++ CrystalGallery/crystal/models.py
@@ -62,7 +62,7 @@ class User(AbstractBaseUser):
 class Listing(models.Model):
     DEFAULT_USER = 1
     name = models.CharField(max_length = 100, blank = False)
-    initial = models.DecimalField(max_digits = 10, decimal_places = 2)
+    initial = models.IntegerField()
     user = models.ForeignKey(User, blank = False, on_delete = models.CASCADE, default = DEFAULT_USER)
     display_picture= models.FileField(upload_to='images/', blank=True)  
     explain = models.TextField(default = '')
